+++ src/example/functions.cpp
@@ -8,8 +8,8 @@
 namespace py = pybind11;
 
 namespace example::functions {
-int add(int i, int j) { return i + j; }
-int sub(int i, int j) { return i - j; }
+double add(double i, double j) { return i + j; }
+double sub(double i, double j) { return i - j; }
 double mul(double i, double j) { return i * j; }
 double div(double i, double j) {
   if (j == 0.0) {
